修改配置文件格式
**复现步骤**

非问题不涉及

**原因和解决方案**

过于复杂的结构对于saltstack渲染配置文件而言会带来额外难度，
而stream或zero的端口又基本上都是一样，因此拆分ip和port的配置

**影响范围**

droplet的启动

**验证方案**

151上droplet成功启动

**涉及分支**

* master

**检查项**

- [ ] 需要更新dep依赖
- [ ] 是共性问题(代码中存在类似问题)
- [x] 编译通过
- [x] 单元测试通过
- [ ] 打包成功(如果有修改Makefile)

/assign @xiangyang

+++ sender/flow_sender.go
@@ -13,7 +13,7 @@ type FlowSender struct {
 	sequence uint32
 }
 
-func NewFlowSender(input queue.QueueReader, ip string, port int) *FlowSender {
+func NewFlowSender(input queue.QueueReader, ip string, port uint16) *FlowSender {
 	return &FlowSender{input, NewZMQBytePusher(ip, port), 1}
 }
